+++ src/Utility/Sitemap.php
@@ -55,8 +55,9 @@ class Sitemap extends SitemapBuilder
         $categories = $table->find('active')
             ->select(['id', 'slug'])
             ->contain(['Pages' => function ($query) use ($table) {
-                $query->select(['category_id', 'slug', 'modified']);
-                $query->order([sprintf('%s.modified', $table->Pages->getAlias()) => 'DESC']);
+                $query->find('active')
+                    ->select(['category_id', 'slug', 'modified'])
+                    ->order([sprintf('%s.modified', $table->Pages->getAlias()) => 'DESC']);
 
                 return $query;
             }]);
@@ -107,8 +108,9 @@ class Sitemap extends SitemapBuilder
         $albums = $table->find('active')
             ->select(['id', 'slug'])
             ->contain(['Photos' => function ($query) use ($table) {
-                $query->select(['id', 'album_id', 'modified']);
-                $query->order([sprintf('%s.modified', $table->Photos->getAlias()) => 'DESC']);
+                $query->find('active')
+                    ->select(['id', 'album_id', 'modified'])
+                    ->order([sprintf('%s.modified', $table->Photos->getAlias()) => 'DESC']);
 
                 return $query;
             }]);
@@ -164,8 +166,9 @@ class Sitemap extends SitemapBuilder
         $categories = $table->find('active')
             ->select(['id', 'slug'])
             ->contain(['Posts' => function ($query) use ($table) {
-                $query->select(['category_id', 'slug', 'modified']);
-                $query->order([sprintf('%s.modified', $table->Posts->getAlias()) => 'DESC']);
+                $query->find('active')
+                    ->select(['category_id', 'slug', 'modified'])
+                    ->order([sprintf('%s.modified', $table->Posts->getAlias()) => 'DESC']);
 
                 return $query;
             }]);