+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.17.2
 * added "pending" find method;
 * improved `isAuthorized()` method: any registered user can access public
-    functions; only admin and managers can access all admin actions.
+    functions; only admin and managers can access all admin actions;
+* fixed bug on the sitemap: pending pages, photos and posts are now excluded.
 
 ### 2.17.1
 * fixed bug for `isAuthorized()` method, simple users could not access some